requirements.txt
tqdm>=4.66.1
PyYAML>=6.0.1
requests>=2.31.0